@@ -1,18 +1,11 @@
 import { CommonModule } from "@angular/common";
-import { HttpClientModule } from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import { NgModule } from "@angular/core";
 import { DirectiveModule } from "../directive/diretivas.module";
 import { DadosComponent } from "./dados.component";
 import { DadoService } from "./dados.service";
 
-@NgModule({
-    declarations: [DadosComponent],
-    imports: [
-        CommonModule,
-        HttpClientModule,
-        DirectiveModule,
-    ],
-    exports: [DadosComponent],
-    providers: [DadoService,]
-})
+@NgModule({ declarations: [DadosComponent],
+    exports: [DadosComponent], imports: [CommonModule,
+        DirectiveModule], providers: [DadoService, provideHttpClient(withInterceptorsFromDi()),] })
 export class DadosModule { }
